feat(markdown): share Pikchr diagram renderer
